Fix building with LAPACK
On Debian cblas.h and lapacke.h are not installed in the same include
directory. This does not matter when compiling, but the custom header
check in OpenCV fails on it. Force OpenCV to use the standard CMake
check_include_file function to do the check so it succeeds.

Also add a missing dependency on liblapacke-dev to provide
lapacke.h.