Build highgui module with QT support
Remove the dependency on the removed non-free Milky icon set. Four
directional arrows are replaced by standard Qt icons. The other icons
are replaced by corresponding icons from the freedesktop.org Icon Naming
Specification.

This also enables OpenGL support as was intended from the
-DWITHOPENGL=ON flag that was already in d/rules.

Remove the dependency on libgtk-3-dev. It is not needed by any module
now.

Add an explicit build dependency on libharfbuzz-dev, this was previously
indirectly provided by the dependency on libgtk-3-dev.

Add cvv to built contrib modules. This is now built because its
dependency on Qt is satisfied.

Closes: 958028